Cambio tipo de dato BDD
double a numerical

+++ src/main/java/com/libcentro/demo/utils/command/UpdateProductCommand.java
@@ -34,19 +34,23 @@ public class UpdateProductCommand implements Command {
         boolean precioVentaCambiado = viejoProducto.getPrecio_venta() != nuevoProducto.getPrecio_venta();
 
         if (costoCambiado && cantidadCambiada) {
-            HistorialCosto historialCosto = historialService.crearHistorialCosto (nuevoProducto,nuevoProducto.getCosto_compra(),cantidad, HistorialCosto.Estado.SIGUIENTE);
-            historialService.save (historialCosto);
+            if(Math.abs (nuevoProducto.getCosto_compra ()-viejoProducto.getCosto_compra()) > viejoProducto.getCosto_compra () * 0.0025d) {
+                HistorialCosto historialCosto = historialService.crearHistorialCosto (nuevoProducto, nuevoProducto.getCosto_compra (), cantidad, HistorialCosto.Estado.SIGUIENTE);
+                historialService.save (historialCosto);
+            }
         } else if ( cantidadCambiada ) {
             HistorialCosto historialCosto = historialService.findLastHistorialCosto (viejoProducto);
-
             historialCosto.setCantidad(historialCosto.getCantidad() + cantidad);
             historialService.save(historialCosto);
         } else if( costoCambiado ){
-            throw new RuntimeException("No se puede cambiar el costo sin una variación en la cantidad");
+            HistorialCosto historialCosto = historialService.findLastHistorialCosto (viejoProducto);
+            historialCosto.setCosto_compra (nuevoProducto.getCosto_compra());
         }
         if (precioVentaCambiado) {
-            HistorialPrecio nuevoHistorialPrecio = historialService.crearHistorialPrecio(nuevoProducto,nuevoProducto.getPrecio_venta());
-            historialService.save(nuevoHistorialPrecio);
+            if(Math.abs (nuevoProducto.getPrecio_venta ()-viejoProducto.getPrecio_venta()) > viejoProducto.getPrecio_venta() * 0.0025d) {
+                HistorialPrecio nuevoHistorialPrecio = historialService.crearHistorialPrecio (nuevoProducto, nuevoProducto.getPrecio_venta ());
+                historialService.save (nuevoHistorialPrecio);
+            }
         }
         productoService.saveProducto (nuevoProducto);
 
@@ -56,30 +60,35 @@ public class UpdateProductCommand implements Command {
     public void undo() {
 
         int cantidad = nuevoProducto.getStock() - viejoProducto.getStock();
+
         boolean costoCambiado = viejoProducto.getCosto_compra() != nuevoProducto.getCosto_compra();
         boolean cantidadCambiada = cantidad != 0;
         boolean precioVentaCambiado = viejoProducto.getPrecio_venta() != nuevoProducto.getPrecio_venta();
 
-
-        if (costoCambiado && cantidad > 0) {
-            HistorialCosto ultimoHistorial = historialService.findLastHistorialCosto(nuevoProducto);
-            historialService.delete(ultimoHistorial);
-        } else if (cantidadCambiada && !costoCambiado) {
-            HistorialCosto historialExistente = historialService.findLastHistorialCosto(nuevoProducto);
-            historialExistente.setCantidad(historialExistente.getCantidad() - cantidad);
-            historialService.save(historialExistente);
+        if (costoCambiado && cantidadCambiada) {
+            if (Math.abs(nuevoProducto.getCosto_compra() - viejoProducto.getCosto_compra()) > viejoProducto.getCosto_compra() * 0.0025d) {
+                HistorialCosto ultimoHistorial = historialService.findLastHistorialCosto(nuevoProducto);
+                historialService.delete(ultimoHistorial);
+            }
+        } else if (cantidadCambiada) {
+            HistorialCosto historialCosto = historialService.findLastHistorialCosto(nuevoProducto);
+            historialCosto.setCantidad(historialCosto.getCantidad() - cantidad);
+            historialService.save(historialCosto);
+        } else if (costoCambiado) {
+            HistorialCosto historialCosto = historialService.findLastHistorialCosto(nuevoProducto);
+            historialCosto.setCosto_compra(viejoProducto.getCosto_compra());
         }
 
-
         if (precioVentaCambiado) {
-
-            HistorialPrecio ultimoHistorialPrecio = historialService.findLastHistorialPrecio(nuevoProducto);
-            historialService.delete(ultimoHistorialPrecio);
+            if (Math.abs(nuevoProducto.getPrecio_venta() - viejoProducto.getPrecio_venta()) > viejoProducto.getPrecio_venta() * 0.0025d) {
+                HistorialPrecio ultimoHistorialPrecio = historialService.findLastHistorialPrecio(nuevoProducto);
+                historialService.delete(ultimoHistorialPrecio);
+            }
         }
 
         nuevoProducto.setCosto_compra(viejoProducto.getCosto_compra());
         nuevoProducto.setPrecio_venta(viejoProducto.getPrecio_venta());
         nuevoProducto.setStock(viejoProducto.getStock());
-        productoService.saveProducto (viejoProducto);
+        productoService.saveProducto(viejoProducto);
     }
 }
